library: destroy hash table after client socket has been closed
If the socket is not closed, even during free_client() a receive
callback(recv_cb()) can be invoked by receiving data from socket.
If recv_cb() is invoked after destroying the hash table, the destroyed
hash table is accessed and segfault is occured.
To prevent the segfault, the hash table should be destroyed after
the socket has been closed.

Change-Id: Ifa59608350a281fc8988042482bdd025b8cc6d59
Signed-off-by: Suchang Woo <suchang.woo@samsung.com>